requirements.txt
pyyaml
gitpython
beautifulsoup4
tqdm
requests